Refactor transform graphql schema for single source
